docs:목데이터 수정에 따른 불필요한 파일 삭제
