+++ client/src/component/AlbumTitleList.js
@@ -1,11 +1,14 @@
 import "./AlbumTitleList.css";
 // 앨범 제목 목록을 보여주는 컴포넌트
-function AlbumTitleList({ albumTitles }) {
+function AlbumTitleListTest({ albumTitles }) {
   return (
     <>
       <h3 style={{ marginBottom: "4px" }}>앨범 목록</h3>
       {/* 리스트 스타일 제거: 점 없애고 들여쓰기 제거 */}
-      <ul className="titleList">
+      <ul
+        className="titleList"
+        style={{ display: "flex", flexDirection: "column", height: "120px" }}
+      >
         {/* 앨범 제목들을 하나씩 렌더링 */}
         {albumTitles.map((albumTitle, index) => (
           <li key={index} className="title">
@@ -17,4 +20,4 @@ function AlbumTitleList({ albumTitles }) {
   );
 }
 
-export default AlbumTitleList;
+export default AlbumTitleListTest;
