+++ client/src/page/OurAlbumPage.js
@@ -1,14 +1,14 @@
 import React, { useState, useEffect } from "react";
+import Header from "../component/Header";
+import Footer from "../component/Footer";
 import Container from "../component/Container";
 import Calender from "../component/Calender";
 import AddAlbum from "../component/AddAlbum";
 import CurrentGroup from "../component/CurrentGroup";
 import Groups from "../component/Groups";
 import AlbumList from "../component/AlbumList";
-import getGroup from "../api/getGroup";
-import getGroupAlbums from "../api/getGroupAlbums";
-import Header from "../component/Header";
-import Footer from "../component/Footer";
+import getGroupTest from "../api/getGroupTest";
+import getGroupAlbumsTest from "../api/getGroupAlbumsTest";
 
 function OurAlbumPage() {
   const [groupList, setGroupList] = useState([]); // 그룹명과 해당 그룹 멤버들의 리스트
@@ -17,20 +17,22 @@ function OurAlbumPage() {
   const [albumTitlesByGroup, setAlbumTitlesByGroup] = useState({}); //그룹ID에 대한 앨범 목록 객체
   //초기 그룹 정보, 앨범 가져오기
   useEffect(() => {
-    const groups = getGroup(); // 그룹 데이터 불러오기
-    setGroupList(groups);
+    const groups = getGroupTest(); // 그룹 데이터 불러오기
 
+    setGroupList(groups);
     if (groups.length > 0) {
       const firstGroup = groups[0]; //항상 첫번째 그룹 선택
-      const firstGroupAlbums = getGroupAlbums(firstGroup.id); //항상 첫번째 그룹의 앨범 선택
+      const firstGroupAlbums = getGroupAlbumsTest(firstGroup.group_id); //항상 첫번째 그룹의 앨범 선택
 
-      setSelectedGroupId(firstGroup.id); // 선택된 그룹 ID
+      setSelectedGroupId(firstGroup.group_id); // 선택된 그룹 ID
       setGroupAlbums(firstGroupAlbums); // 선택된 그룹의 앨범 전체 데이터
       const initTitlesByGroup = {};
       groups.forEach((group) => {
         //처음부터 모든 그룹의 앨범 제목 목록을 한 번에 저장
-        const albums = getGroupAlbums(group.id);
-        initTitlesByGroup[group.id] = albums.map((album) => album.title);
+        const albums = getGroupAlbumsTest(group.group_id);
+        initTitlesByGroup[group.group_id] = albums.map(
+          (album) => album.album_name
+        );
       });
       setAlbumTitlesByGroup(initTitlesByGroup);
     }
@@ -39,7 +41,7 @@ function OurAlbumPage() {
   //그룹id가 바뀔 대마다 그룹 앨범 가져오기
   useEffect(() => {
     if (selectedGroupId) {
-      const albums = getGroupAlbums(selectedGroupId); // 그룹 ID로 앨범 가져오기
+      const albums = getGroupAlbumsTest(selectedGroupId); // 그룹 ID로 앨범 가져오기
       setGroupAlbums(albums); // 상태에 저장
     }
   }, [selectedGroupId]);
@@ -54,7 +56,7 @@ function OurAlbumPage() {
       <Header />
       <Container
         style={{
-          margin: "0 auto",
+          margin: "28px auto",
           padding: "0 40px",
           position: "relative",
           height: "2942px",
@@ -62,6 +64,7 @@ function OurAlbumPage() {
         }}
       >
         <Calender
+          type="group"
           groupAlbums={groupAlbums}
           selectedGroupId={selectedGroupId}
           albumTitlesByGroup={albumTitlesByGroup}
@@ -93,7 +96,12 @@ function OurAlbumPage() {
             />
             <div>
               {/*그룹별 앨범 목록을 보여주는 컴포넌트*/}
-              <AlbumList albums={groupAlbums} />
+              <AlbumList
+                albums={groupAlbums}
+                type="group"
+                selectedGroupId={selectedGroupId}
+                basePath="/our-album"
+              />
             </div>
           </div>
         </div>
